% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertID_HNGC2UniProt.R
\name{convertID_HGNC2UniProt}
\alias{convertID_HGNC2UniProt}
\title{Convert HGNC gene symbol to UniProt Accession IDs}
\usage{
convertID_HGNC2UniProt(HGNCSymbol)
}
\arguments{
\item{HGNCSymbol}{HUGO gene symbol}
}
\value{
UniProt Accession ID
}
\description{
Convert HGNC gene symbol to UniProt Accession IDs using BioMart
}

