% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertID_UniProt2HNGC.R
\name{convertID_UniProt2HGNC}
\alias{convertID_UniProt2HGNC}
\title{Convert HGNC gene symbol to UniProt Accession IDs}
\usage{
convertID_UniProt2HGNC(UniProtID)
}
\arguments{
\item{HGNCSymbol}{HUGO gene symbol}
}
\value{
UniProt Accession ID
}
\description{
Convert HGNC gene symbol to UniProt Accession IDs using BioMart
}

