% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retrieveDomains.R
\name{retrieveDomains}
\alias{retrieveDomains}
\title{Retrieve conserved domains}
\usage{
retrieveDomains(UniProtID)
}
\arguments{
\item{UniProtID}{UniProt Accession ID}
}
\value{
conserved domains
}
\description{
Retrieve conserved domains
}

