% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retrieveProtLen.R
\name{retrieveProtLen}
\alias{retrieveProtLen}
\title{Retrieve protein lengths}
\usage{
retrieveProtLen(UniProtID)
}
\arguments{
\item{UniProtID}{UniProt Accession ID}
}
\value{
protein length
}
\description{
Retrieve protein lengths using UniProt web service
}

