% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMutations.R
\name{plotMutations}
\alias{plotMutations}
\title{Plot mutations}
\usage{
plotMutations(x, y, protLen, annotatePos = c(), annotateSymbol = "*")
}
\arguments{
\item{x}{list of mutation positions}

\item{y}{list of mutation counts for each corresponding position mutated}

\item{protLen}{protein length}

\item{annotatePos}{positions to annotate, default empty list}

\item{annotateSymbol}{symbol to use for annotation, default *}
}
\value{
plots
}
\description{
Plots mutations given mutation data, and matched
}

