% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMutations.R
\name{plotMutations}
\alias{plotMutations}
\title{Plot mutations}
\usage{
plotMutations(x, y, UniProtID = NA, protLen = NA, annotatePos = c(),
  annotateSymbol = "*", annotateProt = c(), xaxis = TRUE, yaxis = TRUE,
  xlab = "Amino acid", ylab = "Number of Mutations", main = "[gene]")
}
\arguments{
\item{x}{list of mutation positions}

\item{y}{list of mutation counts for each corresponding position mutated}

\item{protLen}{protein length; can be from retrieveProtLen}

\item{annotatePos}{positions to annotate, default empty list}

\item{annotateSymbol}{symbol to use for annotation, default *}

\item{annotateProt}{protein annotations, e.g. conserved domains; can be from retrieveDomains}

\item{xaxis}{Boolean to turn on/off x-axis}

\item{yaxis}{Boolean to turn on/off y-axis}

\item{xlab}{x-axis label, default = "Amino acid", only show if xaxis = TRUE}

\item{ylab}{y-axis label, default = "Number of Mutations", only show if yaxis = TRUE}

\item{main}{title, value '[gene]' will be replaced by actual gene name (queried using given UniProtID)}
}
\value{
plots
}
\description{
Plots mutations given mutation data, and matched
}

